+++ client/src/views/errors/error-view.js
@@ -5,17 +5,18 @@ import {Container, Row} from "react-bootstrap";
 import {Link} from "react-router-dom";
 import PropTypes from "prop-types";
 
-const ErrorView = (props) => {
+const ErrorView = ({icon, message, notes = ""}) => {
     return <React.Fragment>
         <Container>
             <Row className="vertical-center">
                 <Col className="text-md-left justify-content-center text-center d-sm-flex d-block">
-                    <div className="mr-sm-5 mr-0">{props.icon}</div>
+                    <div className="mr-sm-5 mr-0">{icon}</div>
                     <div>
                         <h1 className="display-4">Oh Noes!</h1>
-                        <h3>{props.message}</h3>
+                        <h3 className="mb-0">{message}</h3>
+                        {notes && <div className="my-1">{notes}</div>}
                         <Link to="/">
-                            <Button variant="danger" className="mx-0 py-3 px-4">
+                            <Button variant="danger" className="mx-0 py-3 px-4 mt-1">
                                 Back to the Main Page
                             </Button>
                         </Link>
@@ -27,8 +28,9 @@ const ErrorView = (props) => {
 };
 
 ErrorView.propTypes = {
-    icon: PropTypes.object,
-    message: PropTypes.string
+    icon: PropTypes.object.isRequired,
+    message: PropTypes.string.isRequired,
+    notes: PropTypes.string
 };
 
 export default ErrorView;