+++ CHANGELOG.md
@@ -17,3 +17,4 @@
 * Comments should load in good order now, before later pages could be returned earlier and would load first
 * Dark mode received some improvements, theme colors are lighten to fit it better and badges are changed as well
 * Server side messages when editing/updating stuff will now return more user friendly messages on failure
+* In case of client app crash, error crash view will be displayed now